+++ elec_en.py
@@ -9,22 +9,22 @@ def end():
   quit
 
 def voltage():
-  r = int(input("Enter resistance value: "))
-  i = int(input("Enter current: "))
+  r = float(input("Enter resistance value: "))
+  i = float(input("Enter current: "))
   u = r * i
   print ("The voltage value is {} Volts" . format(u))
   main()
 
 def resistance():
-  u = int(input("Enter voltage value: "))
-  i = int(input("Enter current: "))
+  u = float(input("Enter voltage value: "))
+  i = float(input("Enter current: "))
   r = u / i
   print("The resistance value is  {} Ohms" . format(r))
   main()
 
 def current():
-  u = int(input("Enter voltage value: "))
-  r = int(input("Enter resistance value: "))
+  u = float(input("Enter voltage value: "))
+  r = float(input("Enter resistance value: "))
   i = u / r
   print("The current is {} amps" . format(i))
   main()
